modularización de publications
