+++ src/componentes/newAccount.js
@@ -108,10 +108,10 @@ export const newAccount = (navigateTo) => {
     try {
       inputsFormats(inputEmail, inputPassword); // valida que las entradas sean correctas...
       currentUser = await createUser(inputEmail.value, inputPassword.value);
-      console.log(currentUser);
+      // console.log(currentUser);
       await insertNewUserDB(inputNickname.value, currentUser.uid, allUsers);
       // Crea el usuario e ingresa
-      // navigateTo('/publications'); // Se mueve a la vista de publicaciones
+      navigateTo('/publications'); // Se mueve a la vista de publicaciones
       return currentUser;
     } catch (e) {
       errorInvalidPassword.innerText = e.message;
